Fixed two collectibles.py
